+++ server/models/books.js
@@ -1,6 +1,4 @@
-const { router } = require('./index.js');
 const mongoose = require('./mongoose.js');
-
 const Schema = mongoose.Schema;
 
 const booksSchema = new Schema({
@@ -11,11 +9,5 @@ const booksSchema = new Schema({
 });
 
 const Books = mongoose.model('Books', booksSchema);
-
-router.get('/books', (req, res) => {
-  Books.find().then(books => {
-    res.json(books);
-  });
-});
 
 module.exports = Books;
