+++ server/routes/user.js
@@ -1,0 +1,5 @@
+const { user } = require('./routes');
+
+user.get('/profile', function(req, res) {
+  res.send(req.user);
+});
