+++ src/navbar/Navbar.jsx
@@ -14,6 +14,12 @@ function Navbar() {
         <li>
           <Link to="/login">Login</Link>
         </li>
+        <li>
+          <Link to="/user/profile">Profile</Link>
+        </li>
+        <li>
+          <Link to="/logout">Logout</Link>
+        </li>
       </ul>
     </div>
   );
